backend/shortlistai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.30.1
pdfplumber==0.11.1
python-docx==1.1.2
numpy==1.26.4
pandas==2.2.2
aiofiles==23.2.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
